Prototype V1

+++ src/main/java/com/laptev/controller/OrderController.java
@@ -8,10 +8,12 @@ import com.laptev.service.UserService;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import java.util.Collections;
+import java.util.Optional;
 
 @Controller
 @AllArgsConstructor
@@ -24,6 +26,12 @@ public class OrderController {
 
     private final RequestService requestService;
 
+    @GetMapping("/allOrders")
+    public String allOrders(Model model) {
+        model.addAttribute("allOrders", orderService.allOrders());
+        return "orders";
+    }
+
     @PostMapping("/creatingOrder")
     public String processingRoom(@RequestParam(required = true, defaultValue = "") Long roomNumber,
                                  @RequestParam(required = true, defaultValue = "") Long userId,
@@ -34,13 +42,31 @@ public class OrderController {
         Order order = new Order();
         order.setUser(userService.findUserById(userId));
         order.setDurationHours(durationHours);
-        order.setRooms(Collections.singleton(roomService.findById(roomNumber)));
+        order.setRoom(roomService.findById(roomNumber));
         order.setPlaces(places);
         orderService.saveOrder(order);
-        requestService.deleteRequest(requestId);
-//        requestService.updateRequestStatus(requestService.findById(requestId), RequestStatus.COMPLETED);
+        requestService.updateRequestStatus(requestService.findById(requestId), RequestStatus.COMPLETED);
         roomService.updateRoomStatus(roomService.findById(roomNumber), RoomStatus.BOOKED);
-
+        orderService.sendMail(userService.findUserById(userId).getEmail(), requestService.findById(requestId), order);
+//        requestService.deleteRequest(requestId);
         return "redirect:/adminRequest";
     }
+
+    @GetMapping("/confirmRoom")
+    String confirmRoom(@RequestParam("confirm") String confirm,
+                       @RequestParam("orderId") Long orderId,
+                       @RequestParam("requestId") Long requestId) {
+
+        if (confirm.equals("yes")) {
+            orderService.updateOrderStatus(orderService.findById(orderId), OrderStatus.CONFIRM);
+            requestService.deleteRequest(requestId);
+
+        } else if (confirm.equals("no")) {
+            requestService.updateRequestStatus(requestService.findById(requestId), RequestStatus.IN_PROGRESS);
+            roomService.updateRoomStatus(roomService.findById(orderService.findById(orderId).getRoom().getRoomNumber()), RoomStatus.EMPTY);
+            orderService.deleteOrderById(orderId);
+        }
+        return "redirect:/";
+    }
 }
+// посмотреть как я сделал так, чтобы BOOKED комнаты не показывались в списке подходящих
